requirements.txt
pandas
streamlit==2.2.2
plotly
python-dotenv
altair
bokeh
requests
